+++ lib/rspec/wait/error.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module RSpec
-  module Wait
-    Error = Class.new(StandardError)
-    TimeoutError = Class.new(Error)
-  end
-end